test_model.py
import unittest
import tempfile
import os
from transcription_model import TranscriptionModel

class TestTranscriptionModel(unittest.TestCase):
    def setUp(self):
        self.transcription_model = TranscriptionModel()

    def test_load_audio(self):
        # Create a temporary audio file
        with tempfile.NamedTemporaryFile(suffix='.mp3') as f:
            file_path = f.name
            f.write(b'test audio data')

            # Ensure that the file can be loaded
            self.transcription_model.load_audio(file_path)

            # Ensure that an invalid file path raises a ValueError
            with self.assertRaises(ValueError):
                self.transcription_model.load_audio('invalid_path')

            # Ensure that a file with an invalid extension raises a ValueError
            with tempfile.NamedTemporaryFile(suffix='.txt') as f:
                invalid_file_path = f.name
                with self.assertRaises(ValueError):
                    self.transcription_model.load_audio(invalid_file_path)

    def test_transcribe_audio(self):
        # TODO: Write test cases for the transcribe_audio method
        pass

    def test_manipulate_text(self):
        # TODO: Write test cases for the manipulate_text method
        pass

    def test_save_transcript(self):
        # Create a temporary transcript file
        with tempfile.NamedTemporaryFile(suffix='.txt') as f:
            file_path = f.name
            transcript = 'test transcript'

            # Save the transcript to the file
            self.transcription_model.save_transcript(transcript, file_path)

            # Ensure that the file was created and contains the transcript
            with open(file_path, 'r') as f:
                self.assertEqual(f.read(), transcript)

    def tearDown(self):
        # Remove any temporary files created during testing
        for file_path in self.transcription_model.audio_files + self.transcription_model.transcript_files:
            if os.path.exists(file_path):
                os.remove(file_path)

if __name__ == '__main__':
    unittest.main()
